% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedfinaldegree.R
\name{adjustedfinaldegree}
\alias{adjustedfinaldegree}
\title{Adjusted Final Degree Classification}
\usage{
adjustedfinaldegree(score)
}
\arguments{
\item{score}{it's one possible mark between 18 and 30 that you can take in your missing exams}
}
\value{
adjusted final degree classification
}
\description{
Takes a possible score to all the exams that you still need to do and calculates your "adjusted" final degree
}
