% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradeCalculatorGUI.R
\name{GradeCalculatorGUI}
\alias{GradeCalculatorGUI}
\title{GradeCalculatorGUI}
\usage{
GradeCalculatorGUI()
}
\value{
shiny app
}
\description{
Runs a Graphical User Interface to calculate your final degree
}
