% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{File DSE}
\format{data frame}
\source{
\url{https://github.com/unimi-dse/c34a5d0e/tree/master/inst/extdata/FileDse.csv}
}
\usage{
mydata
}
\description{
A dataset containing few students' name, gender, age,	nationality
from the Master in Data Science and Economics, University of Milan.
This dataset contains also taken exams (subjects, score, date and cfu) for each student.
}
\keyword{datasets}
