% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDensity.R
\name{ScoreDensity}
\alias{ScoreDensity}
\title{Score Density}
\usage{
ScoreDensity()
}
\value{
plot of the Score density
}
\description{
This function reads the data into the dataframe and calculates how many times occur a certain score
}
