% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDensityForStudent.R
\name{ScoreDensityForStudent}
\alias{ScoreDensityForStudent}
\title{ScoreDensityForStudent}
\usage{
ScoreDensityForStudent()
}
\value{
plot of the score density for every student
}
\description{
This function reads the data into the dataframe and calculates the mean, the variance, the min/max score for every student
}
