% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFrequency.R
\name{TermFrequency}
\alias{TermFrequency}
\title{TermFrequency}
\usage{
TermFrequency()
}
\value{
plot of the density of exams in a term
}
\description{
This function reads the data into the dataframe how many exams have been in a term
}
