% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedaverage.R
\name{weightedaverage}
\alias{weightedaverage}
\title{Weighted Average}
\usage{
weightedaverage(Students)
}
\arguments{
\item{Students}{it's the list of students which you want to calculate the final degree classification}
}
\value{
final degree classification
}
\description{
Takes the list of students, calculates weighted average of all the courses of the students and computes the final degree classification
}
