% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedaverageGUI.R
\name{weightedaverageGUI}
\alias{weightedaverageGUI}
\title{Weighted Average GUI}
\usage{
weightedaverageGUI()
}
\value{
the trend of the final degree classification
}
\description{
This function reads the data into the dataframe and give us the trend of the final grades
}
